v0.2.5.012 — Phase 4.x：agent 動態切 character voice + 修 minimax speed cast 為 float
Phase 4.x 完整版：
- realtime_agent: 從 char_ctx.voice_id_minimax 拿角色自己的 voice，沒設 fallback default (吉娜)
- realtime_agent: 從 char_ctx.tts_settings_minimax 讀 speed/pitch，沒設用 1.0/0
- 聖嚴 (speed=0.9 pitch=-1)、吉娜 (speed=1.2)、馬雲等都會用各自 voice + 各自節奏

順手修 minimax_tts.py：
- 江彬原寫 speed/vol 用 int()，但 MiniMax API speed/vol 是 float (0.5-2.0)
- 0.9 / 1.2 cast int 變 0/1 → 念太慢或不發聲
- 改 float()，pitch 維持 int (-12 ~ 12)

Cloud Run revision 00011 已部署。聖嚴撥號這次應該聽到聖嚴本人聲音。

+++ agent/minimax_tts.py
@@ -120,9 +120,11 @@ class MiniMaxChunkedStream(tts.ChunkedStream):
             "stream": False,
             "voice_setting": {
                 "voice_id": self._opts.voice_id,
-                "speed": int(self._opts.speed),
-                "vol": int(self._opts.vol),
-                "pitch": int(self._opts.pitch),
+                # 江彬 minimax_tts 原寫 int()，但 MiniMax API speed/vol 是 float（0.5-2.0），
+                # 0.9 / 1.2 這種小數 cast 成 0/1 → 念太慢或不發聲。修正為 float。
+                "speed": float(self._opts.speed),
+                "vol": float(self._opts.vol),
+                "pitch": int(self._opts.pitch),  # pitch 是 int (-12 ~ 12) 不變
             },
             "audio_setting": {
                 "sample_rate": int(self._opts.sample_rate),
